Display a simple message when there are no tasks.
When there are no tasks, the worker script will simply sleep and wait until there is a task. However, no output is generated when this happens, so the user may think an error has occurred. This change adds a message so that the user knows the worker is waiting for a task.

Note that as implemented, the worker will display the message repeatedly if it needs to wait for an extended period of time.